+++ GR_Init.lua
@@ -5,7 +5,8 @@ ICON_PATH = 'Interface\\AddOns\\GuildRecruiter\\Images\\'
 
 GR = LibStub('AceAddon-3.0'):NewAddon('GuildRecruiter', 'AceConsole-3.0', 'AceEvent-3.0', 'AceHook-3.0', 'AceComm-3.0')
 GR.debug = false
-GR.isBeta = true
+GR.isTest = true
+GR.testLevel = 'Alpha'
 GR.author  = C_AddOns.GetAddOnMetadata('GuildRecruiter', 'Author')
 GR.version = C_AddOns.GetAddOnMetadata('GuildRecruiter', 'Version')