+++ CHANGELOG.md
@@ -5,4 +5,5 @@ Features:
 - Reworked Settings and the GM/Personal invite system should be more clear.
 - Black List will now immediately remove a player from the black list and there will not be a wait time.
 - Added the ability to add zones to the zone list.  This will require a bit of searching on the internet to find the zone ID.
+- Added Shift+Click to directly open scanner.